move constants to a separate module
